// 3ME_UI/src/forgotCredentials/forgotPassword/forgotPasswordContainer/forgotCredentialPages.enum.ts
export enum ForgotCredentialPages {
  ENTER_USERNAME = "enterusername",
  CODE_VERIFICATION = "codeverification",
  PASSWORD_RESET_PAGE = "passwordReset",
}

export enum ForgotPasswordOption {
  EMAIL = 0,
  SMS = 1,
  NONE = 9999,
}

export enum ForgotPasswordChannel {
  EMAIL = 0,
  SMS = 1,
  NONE = 9999,
}